pyproject.toml
[tool.poetry]
name = "limits"
version = "0.2.0"
description = "Vino Bank demo limits management module"
authors = ["Li Lin <guru.lin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aws-lambda-powertools = "^1.24.2"
pytest = "^6.2.5"
pytest-mock = "^3.5.1"
pytest-cov = "^3.0.0"
requests = "^2.27.1"
ulid-py = "^1.1.0"
pycognito = "^2022.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
# formatting, linting
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-black = "^0.2.3"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.7.0"
flake8-debugger = "^4.0.0"
flake8-fixme = "^1.1.1"
flake8-isort = "^4.1.1"
flake8-variables-names = "^0.0.4"
isort = "^5.10.1"
pre-commit = "^2.17.0"
# mypy type checking
mypy = "^0.931"
types-requests = "^2.27.8"
types-setuptools = "^57.4.8"
types-toml = "^0.10.3"
boto3-stubs = "1.21.2"

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["src", "test"]

[tool.black]
line-length = 120

[tool.pytest.ini_options]
minversion = "6.0"
# addopts = "--cov=limits --cov-report html --cov-report xml"
testpaths = [
    "tests/unit",
]

[tool.coverage.run]
source = ["limits"]
omit = ["tests/*"]

[tool.coverage.html]
directory = "test_report"
